Moved ChaninedStreaming to a separate file and renamed to ChainedBroadcast.
